% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS mapper}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A integer representing a state in the FARS data}

\item{year}{A integer specifying the year.}
}
\value{
If the data for the year is available and the state.num is valid
 the output will be a plot with an outline of the state with points showing the
 locations of accidents during that year
 If an invalid state.num is passed a an argument the following message will be returned
 "invalid STATE number: state.num"
 If there are no accidents for the year and state specified the following message will
 be returned "no accidents to plot"
}
\description{
\code{fars_map_state} plots the number of accidents per state per year specified in the arguments
}
\note{
requires maps and graphics packages
}
\examples{
fars_map_state('10',2013)
fars_map_state('01',2015)
}
