% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{FARS reader}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
data A data frame tbl
}
\description{
\code{fars_read} returns a data frame tbl sourced from the file in its arguments
}
\note{
requires readr and dplyr packages

If the files does not exist will return an error message "file '", filename, "' does not exist"
}
\examples{
fars_read("accident.2014.csv.bz2")
fars_read("xyz.csv")
}
