% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{FARS reader by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numbers representing years}
}
\value{
a data.frame with with the month and year as columns
}
\description{
\code{fars_read_years} reads files into a data.frame with two columns MONTH and year
}
\note{
requires dplyr package
}
\examples{
fars_read(2014)
#'fars_read(2014:2015)
}
