% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Filename maker}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric.}
}
\value{
sprintf A string
}
\description{
\code{make_filename} returns a string containg the file name for the year specified in the arguments
}
\note{
year must be compatible with as.integer
}
\examples{
fars_read(2014)
}
